+++ R/gg_overview.R
@@ -1,0 +1,109 @@
+#' Plot an overview of dataset intervals with implicit missing data
+#'
+#' @inheritParams gg_day
+#' @param Datetime.column column name that contains the datetime. Defaults to
+#'   `"Datetime"` which is automatically correct for data imported with
+#'   [LightLogR]. Expects a `symbol`. Needs to be part of the `dataset`.
+#' @param Id.column The column name of the Id column (default is `Id`), needs to
+#'   be in the `dataset`. This is also used as the y-axis variable and is the
+#'   minimum grouping variable.
+#' @param gap_range_data Optionally provide a `tibble` with `start` and `end`
+#'   `Datetimes` of gaps per group. If not provided, the function uses
+#'   [gap_finder()] to calculate implicit missing data. This might be
+#'   computationally intensive for large datasets and many missing data. In
+#'   these cases it can make sense to calculate those gaps beforehand and
+#'   provide them to the function.
+#' @param ... Additional arguments given to the main [ggplot2::aes()] used for
+#'   styling depending on data within the `dataset`
+#'
+#' @return A `ggplot` object
+#' @export
+#'
+#' @examples
+#' sample.data.environment %>% gg_overview(Id.column = Source)
+gg_overview <- function(dataset,
+                        Datetime.column = Datetime,
+                        Id.column = Id,
+                        gap_range_data = NULL,
+                        ...,
+                        interactive = FALSE) {
+  
+  # Initial Checks ----------------------------------------------------------
+  
+  #dataset must be a dataset with a posixct datetime column
+  
+  
+  # Function ----------------------------------------------------------
+  
+  #make sure the dataset is at least grouped by the Id.column
+  dataset <- dataset %>% dplyr::group_by({{ Id.column }}, .add = TRUE)
+  
+  #calculate the ranges of gap data if none is provided
+  if(is.null(gap_range_data)) {
+    gap_range_data <- 
+      dataset %>% 
+      gap_finder(gap.data = TRUE, silent = TRUE) %>% 
+      dplyr::group_by(gap.id, .add = TRUE) %>% 
+      dplyr::filter(dplyr::n() > 1)
+    
+    there_are_gaps <- nrow(gap_range_data) > 0
+    
+    if(there_are_gaps) {
+      gap_range_data <- 
+        gap_range_data %>% 
+        dplyr::summarize(
+          start = min({{ Datetime.column}}), end = max({{ Datetime.column}}))
+    }
+  }
+  
+  #are there missing data
+  there_are_gaps <- nrow(gap_range_data) > 0
+  
+  #only add missing data when there is some
+  implicit_data <- 
+    if(there_are_gaps) {
+      list(
+        ggplot2::geom_linerange(
+          data = gap_range_data,
+          ggplot2::aes(xmin = start, xmax = end, y = {{ Id.column}}),
+          col = "grey", size = 1),
+      ggplot2::labs(
+        caption = 
+          "<b>Dataset intervals</b>, with times of <b style = 'color:grey'>implicit missing data </b>."
+      ))
+    } else list()
+  
+  #calculate the ranges within the dataset per grouping
+  dataset_range <- 
+    dataset %>%
+    dplyr::summarize(
+      start = min({{ Datetime.column}}), end = max({{ Datetime.column}}))
+  
+  Plot <- 
+    #setup the plot
+    dataset_range %>% 
+    ggplot2::ggplot(
+      ggplot2::aes(xmin = start, xmax = end, y = {{ Id.column }},...)
+      ) + 
+    #add the lineranges
+    ggplot2::geom_linerange(size = 1) +
+    implicit_data +
+    #add start and stop points
+    ggplot2::geom_point(
+      ggplot2::aes(x = start, y = {{ Id.column}}), size = 2) +
+    ggplot2::geom_point(
+      ggplot2::aes(x = end, y = {{ Id.column}}), size = 2) +
+    #general information
+    ggplot2::labs(x = "Datetime")+
+    #theming and styling
+    cowplot::theme_cowplot()+
+    ggplot2::theme(
+      panel.grid.major = ggplot2::element_line(colour = "grey98"), 
+      plot.caption = ggtext::element_markdown())
+  
+  # Return ----------------------------------------------------------
+  if(interactive) {
+    Plot %>% plotly::ggplotly()
+  } else Plot
+  
+}