+++ NEWS.md
@@ -1,5 +1,7 @@
 # LightLogR 0.2.1.9000
 
+* Added the `gg_overview()` function to get a sense for the timeframe of measurement data.
+
 * Added the family of `regularize` functions to find and deal with implicit missing data. These functions include `dominant_epoch()`, `gapless_Datetimes()`, `gap_handler()`, and `gap_finder()`. 
 
 * A ton of updates to documentation, unit tests, and bug fixes.